# Minor interpreter performance improvements