@shlevy 2018-02-17
@dezgeg Thanks! Would you mind showing per-commit profiling? Or is there some reason to think these are synergistic?