@dezgeg 2018-02-19
They are:
````
3.256 (baseline) | Merge pull request #1873 from matthewbauer/fix-manpath
3.253  -0.11 %   | libexpr: Pre-reserve space in string in unescapeStr()
3.232  -0.64 %   | libexpr: Optimize prim_attrNames a bit
3.224  -0.27 %   | libexpr: Rely on Boehm returning zeroed memory in EvalState::allocEnv()
3.219  -0.16 %   | libexpr: Optimize prim_derivationStrict by using more symbol comparisons
3.221  +0.06 %   | libexpr: Remove unnecessary drvName assignment in prim_derivationStrict
3.202  -0.57 %   | libexpr: Avoid an unnecessary string copy in prim_derivationStrict
````

Which have the roughly same orders of magnitude except with more random variation (especially on the "libexpr: Remove unnecessary drvName assignment in prim_derivationStrict" one).